Add VulnDBHQ::PrivatePage class and getter
The ::PrivatePage class inherits from ::Base which holds the identity map and
some other helper features for instance to create an instance from the HTTP
body response from the server.

There are some methods in the ::Client to retrieve a PrivatePage (using an :id)
or to get a collection of them (not yet implemented).